growing-mesh/solver/requirements.txt
mpi4py
pyprecice @ git+https://github.com/precice/python-bindings.git@develop
